database: Share `DataSource` code
