+++ audio-api/src/main/scala/no/ndla/audioapi/ComponentRegistry.scala
@@ -10,7 +10,6 @@ package no.ndla.audioapi
 
 import com.zaxxer.hikari.HikariDataSource
 import no.ndla.audioapi.controller.*
-import no.ndla.audioapi.integration.*
 import no.ndla.audioapi.model.api.ErrorHandling
 import no.ndla.audioapi.repository.{AudioRepository, SeriesRepository}
 import no.ndla.audioapi.service.*
@@ -18,6 +17,7 @@ import no.ndla.audioapi.service.search.*
 import no.ndla.common.Clock
 import no.ndla.common.aws.NdlaS3Client
 import no.ndla.common.configuration.BaseComponentRegistry
+import no.ndla.database.DataSource
 import no.ndla.network.NdlaClient
 import no.ndla.network.tapir.TapirApplication
 import no.ndla.search.{BaseIndexService, Elastic4sClient}
