+++ audio-api/src/main/scala/no/ndla/audioapi/model/api/Error.scala
@@ -9,9 +9,9 @@
 package no.ndla.audioapi.model.api
 
 import no.ndla.audioapi.Props
-import no.ndla.audioapi.integration.DataSource
 import no.ndla.common.Clock
 import no.ndla.common.errors.{AccessDeniedException, FileTooBigException, ValidationException}
+import no.ndla.database.DataSource
 import no.ndla.network.model.HttpRequestException
 import no.ndla.network.tapir.{AllErrors, ErrorBody, TapirErrorHandling, ValidationErrorBody}
 import no.ndla.search.NdlaSearchException
