+++ audio-api/src/main/scala/no/ndla/audioapi/repository/AudioRepository.scala
@@ -10,10 +10,10 @@ package no.ndla.audioapi.repository
 
 import com.typesafe.scalalogging.StrictLogging
 import no.ndla.audioapi.Props
-import no.ndla.audioapi.integration.DataSource
 import no.ndla.audioapi.model.api.ErrorHandling
 import no.ndla.audioapi.model.domain.{AudioMetaInformation, Series}
 import no.ndla.common.CirceUtil
+import no.ndla.database.DataSource
 import org.postgresql.util.PGobject
 import scalikejdbc.*
