+++ concept-api/src/main/scala/no/ndla/conceptapi/repository/PublishedConceptRepository.scala
@@ -11,9 +11,9 @@ import com.typesafe.scalalogging.StrictLogging
 import no.ndla.common.CirceUtil
 import no.ndla.common.model.domain.Tag
 import no.ndla.common.model.domain.concept.Concept
-import no.ndla.conceptapi.integration.DataSource
 import no.ndla.conceptapi.model.api.NotFoundException
 import no.ndla.conceptapi.model.domain.{DBConcept, PublishedConcept}
+import no.ndla.database.DataSource
 import org.postgresql.util.PGobject
 import scalikejdbc.*
