+++ database/src/main/scala/no/ndla/database/DatabaseProps.scala
@@ -1,0 +1,28 @@
+/*
+ * Part of NDLA database
+ * Copyright (C) 2024 NDLA
+ *
+ * See LICENSE
+ *
+ */
+
+package no.ndla.database
+
+import no.ndla.common.Environment.prop
+import no.ndla.common.secrets.PropertyKeys
+
+import scala.util.Properties.propOrElse
+
+trait DatabaseProps {
+  def MetaUserName: String    = prop(PropertyKeys.MetaUserNameKey)
+  def MetaPassword: String    = prop(PropertyKeys.MetaPasswordKey)
+  def MetaResource: String    = prop(PropertyKeys.MetaResourceKey)
+  def MetaServer: String      = prop(PropertyKeys.MetaServerKey)
+  def MetaPort: Int           = prop(PropertyKeys.MetaPortKey).toInt
+  def MetaSchema: String      = prop(PropertyKeys.MetaSchemaKey)
+  def MetaMaxConnections: Int = propOrElse(PropertyKeys.MetaMaxConnections, "10").toInt
+}
+
+trait HasDatabaseProps {
+  val props: DatabaseProps
+}
