+++ draft-api/src/main/scala/no/ndla/draftapi/model/api/NDLAErrors.scala
@@ -10,8 +10,8 @@ package no.ndla.draftapi.model.api
 import com.typesafe.scalalogging.StrictLogging
 import no.ndla.common.Clock
 import no.ndla.common.errors.{AccessDeniedException, FileTooBigException, ValidationException}
+import no.ndla.database.DataSource
 import no.ndla.draftapi.Props
-import no.ndla.draftapi.integration.DataSource
 import no.ndla.network.model.HttpRequestException
 import no.ndla.network.tapir.{AllErrors, ErrorBody, TapirErrorHandling}
 import no.ndla.search.{IndexNotFoundException, NdlaSearchException}
