+++ frontpage-api/src/main/scala/no/ndla/frontpageapi/FrontpageApiProperties.scala
@@ -10,32 +10,21 @@ package no.ndla.frontpageapi
 import no.ndla.common.Environment.prop
 import no.ndla.common.configuration.{BaseProps, HasBaseProps}
 import no.ndla.network.{AuthUser, Domains}
-import no.ndla.common.secrets.PropertyKeys
+import no.ndla.database.{DatabaseProps, HasDatabaseProps}
 
-import scala.util.Properties._
+import scala.util.Properties.*
 
-trait Props extends HasBaseProps {
+trait Props extends HasBaseProps with HasDatabaseProps {
   val props: FrontpageApiProperties
 }
 
-class FrontpageApiProperties extends BaseProps {
+class FrontpageApiProperties extends BaseProps with DatabaseProps {
   def ApplicationName            = "frontpage-api"
   val ApplicationPort: Int       = propOrElse("APPLICATION_PORT", "80").toInt
-  val NumThreads: Int            = propOrElse("NUM_THREADS", "200").toInt
   val DefaultLanguage: String    = propOrElse("DEFAULT_LANGUAGE", "nb")
   val Auth0LoginEndpoint: String = s"https://${AuthUser.getAuth0HostForEnv(Environment)}/authorize"
 
-  val SecretsFile = "frontpage-api.secrets"
-
-  def MetaUserName: String    = prop(PropertyKeys.MetaUserNameKey)
-  def MetaPassword: String    = prop(PropertyKeys.MetaPasswordKey)
-  def MetaResource: String    = prop(PropertyKeys.MetaResourceKey)
-  def MetaServer: String      = prop(PropertyKeys.MetaServerKey)
-  def MetaPort: Int           = prop(PropertyKeys.MetaPortKey).toInt
-  def MetaSchema: String      = prop(PropertyKeys.MetaSchemaKey)
-  def MetaMaxConnections: Int = propOrElse(PropertyKeys.MetaMaxConnections, "10").toInt
-  val DefaultPageSize         = 10
-
+  val DefaultPageSize        = 10
   val Domain: String         = propOrElse("BACKEND_API_DOMAIN", Domains.get(Environment))
   val RawImageApiUrl: String = s"$Domain/image-api/raw"
