+++ image-api/src/test/scala/no/ndla/imageapi/TestEnvironment.scala
@@ -11,6 +11,7 @@ package no.ndla.imageapi
 import com.zaxxer.hikari.HikariDataSource
 import no.ndla.common.Clock
 import no.ndla.common.aws.NdlaS3Client
+import no.ndla.database.DataSource
 import no.ndla.imageapi.controller.{
   BaseImageController,
   HealthController,
@@ -19,7 +20,6 @@ import no.ndla.imageapi.controller.{
   InternController,
   RawController
 }
-import no.ndla.imageapi.integration.*
 import no.ndla.imageapi.model.api.ErrorHandling
 import no.ndla.imageapi.repository.*
 import no.ndla.imageapi.service.*
